set env-var PYTHON_BIN to compile uwsgi with correct python-version
with the commit https://github.com/unbit/uwsgi/commit/d584857c781593d6c1f7f2a98503fcdce2ed9893
uwsgi is able handle different python-binaries in the Makefile so its
possible to set the env-var PYTHON_BIN to the desired python-version.
Now we like to compile uwsgi with the python-version buildout is
running (otherwise its possible compile uwsgi with python2.6 because it's
your system default and now wit i.e. python2.7 buildout is running with)
In this commit we set the env-var PYTHON_BIN to currently running
python-bin with sys.executable